+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Fixed 🐞
 
+-   Selected Building now stays highlighted when map is rotated #1498
+
 ### Chore 👨‍💻 👩‍💻
 
 ## [1.62.0] - 2020-11-12